move useDarkTheme out of store methods to avoid hook error, test light mode
