+++ client/App.tsx
@@ -2,14 +2,14 @@ import { Navigate, Route } from 'react-router-native';
 import { Home } from './src/Home/Home';
 import { LoginPage } from './src/Login/LoginPage';
 import { LogMeal } from './src/Meal/Log/LogMeal';
-import { Meal } from './src/Meal/Meal';
-import { Providers } from './src/common/Providers';
+import { Meal } from './src/Meal/MealPage/Meal';
+import { ProvidersWrapper } from './src/common/Providers/Providers';
 import { WTRoutes } from './src/routes';
 import { _isLoggedIn } from './src/utils/storeMethods';
 
 const App = (): JSX.Element => {
   return (
-    <Providers>
+    <ProvidersWrapper>
       <Route
         path={WTRoutes.Entry}
         element={
@@ -24,7 +24,7 @@ const App = (): JSX.Element => {
       <Route path={WTRoutes.Home} element={<Home />} />
       <Route path={WTRoutes.Meal} element={<Meal />} />
       <Route path={WTRoutes.LogMeal} element={<LogMeal />} />
-    </Providers>
+    </ProvidersWrapper>
   );
 };
