+++ client/src/common/Error/Error.tsx
@@ -1,7 +1,7 @@
 import { VStack } from '@react-native-material/core';
 import { Avatar, Text } from 'react-native-paper';
 
-const DEFAULT_ERROR_MSG = `Uh oh... looks like we experienced an serious error. 
+const DEFAULT_ERROR_MSG = `Uh oh... looks like we experienced a serious error. 
 Please try again later, sorry!`;
 
 export const ErrorPage: React.FC<{ message?: string }> = ({
