+++ client/src/common/Providers/apolloClient.ts
@@ -1,0 +1,33 @@
+import { ApolloClient, HttpLink, InMemoryCache, split } from '@apollo/client';
+import { getMainDefinition } from '@apollo/client/utilities';
+import { WebSocketLink } from '@apollo/link-ws';
+
+export const setupApolloClient = () => {
+  const wsLink = new WebSocketLink({
+    uri: `ws://localhost:3000/graphql`,
+    options: {
+      reconnect: true,
+    },
+  });
+
+  const httpLink = new HttpLink({
+    uri: 'http://localhost:3000/graphql',
+  });
+
+  const link = split(
+    ({ query }) => {
+      const definition = getMainDefinition(query);
+      return (
+        definition.kind === 'OperationDefinition' &&
+        definition.operation === 'subscription'
+      );
+    },
+    wsLink,
+    httpLink,
+  );
+
+  return new ApolloClient({
+    link,
+    cache: new InMemoryCache(),
+  });
+};
